fixed server errors
