+++ src/components/MyRoutes.js
@@ -4,6 +4,7 @@ import MainPage from '../protected/MainPage';
 import Login from '../auth/Login';
 import Signup from '../auth/Signup';
 import { RequireAuth } from '../auth/authHelpers';
+import Addfreelancer from './Addfreelancer';
 
 const MyRoutes = () => {
   const routes = useRoutes([
@@ -29,7 +30,7 @@ const MyRoutes = () => {
         },
         {
           path: '/addfreelancer',
-          element: <RequireAuth><div>Link to Add Freelancer</div></RequireAuth>,
+          element: <RequireAuth><Addfreelancer /></RequireAuth>,
         },
         {
           path: '/deletefreelancer',
